halfway through refactoring styles - using percentages and external stylesheets
