+++ client/screens/HomeScreen.js
@@ -21,6 +21,7 @@ import * as Google from 'expo-google-app-auth'
 import {useDispatch, useSelector} from 'react-redux'
 import pinsData from '../store/pins' //fake data for now
 import {getPlants, getUserData, setUserData, setPins} from '../store/actions'
+import styles from '../styles/HomeScreenStyles'
 
 const HomeScreen = props => {
   // const [isFetching, setIsFetching] = useState(false)
@@ -263,148 +264,5 @@ function handleHelpPress() {
     'https://docs.expo.io/versions/latest/workflow/up-and-running/#cant-see-your-changes'
   )
 }
-
-const styles = StyleSheet.create({
-  alert: {
-    backgroundColor: 'grey'
-  },
-  welcomeContainer: {
-    marginTop: 30,
-    justifyContent: 'center',
-    alignItems: 'center'
-  },
-  welcomeImage: {
-    width: 80,
-    height: 80,
-    marginLeft: 10,
-    resizeMode: 'contain',
-    alignItems: 'center',
-    justifyContent: 'center'
-  },
-  title: {
-    marginTop: 5,
-    marginBottom: 0,
-    color: '#000000',
-    textAlign: 'center',
-    fontSize: 40,
-    fontFamily: 'yorkten'
-  },
-  titleTwo: {
-    marginTop: 10,
-    color: '#000000',
-    textAlign: 'center',
-    fontSize: 30,
-    fontFamily: 'yorkten'
-  },
-  subtitle: {
-    color: '#C7CAD4',
-    textAlign: 'center',
-    alignSelf: 'stretch',
-    marginRight: 0,
-    marginLeft: 0,
-    marginTop: 5,
-    fontSize: 15,
-    marginBottom: 5
-  },
-  screenText: {
-    color: '#000000',
-    fontFamily: 'yorkten',
-    fontSize: 18,
-    textAlign: 'center'
-  },
-  screenTextBottom: {
-    color: '#000000',
-    fontFamily: 'yorkten',
-    fontSize: 18,
-    textAlign: 'center'
-  },
-  socialContainer: {},
-  socialColumn: {
-    width: '50%' // 50% of container width
-  },
-  socialLinks: {
-    color: '#6CC7BD',
-    fontSize: 18
-  },
-  label: {
-    borderWidth: 1,
-    marginTop: 20,
-    marginBottom: 20
-  },
-  button: {
-    backgroundColor: '#6CC7BD'
-  },
-  container: {
-    alignSelf: 'stretch'
-  },
-  developmentModeText: {
-    marginBottom: 20,
-    color: 'rgba(0,0,0,0.4)',
-    fontSize: 14,
-    lineHeight: 19,
-    textAlign: 'center'
-  },
-  contentContainer: {
-    paddingTop: 30,
-    flex: 1,
-    textAlign: 'center'
-  },
-  homeScreenFilename: {
-    marginVertical: 7
-  },
-  codeHighlightText: {
-    color: 'rgba(96,100,109, 0.8)'
-  },
-  codeHighlightContainer: {
-    backgroundColor: 'rgba(0,0,0,0.05)',
-    borderRadius: 3,
-    paddingHorizontal: 4
-  },
-  getStartedText: {
-    fontSize: 17,
-    color: 'rgba(96,100,109, 1)',
-    lineHeight: 24,
-    textAlign: 'center'
-  },
-  tabBarInfoContainer: {
-    position: 'absolute',
-    bottom: 0,
-    left: 0,
-    right: 0,
-    ...Platform.select({
-      ios: {
-        shadowColor: 'black',
-        shadowOffset: {width: 0, height: -3},
-        shadowOpacity: 0.1,
-        shadowRadius: 3
-      },
-      android: {
-        elevation: 20
-      }
-    }),
-    alignItems: 'center',
-    backgroundColor: '#fbfbfb',
-    paddingVertical: 20
-  },
-  tabBarInfoText: {
-    fontSize: 17,
-    color: 'rgba(96,100,109, 1)',
-    textAlign: 'center'
-  },
-  navigationFilename: {
-    marginTop: 5
-  },
-  helpContainer: {
-    marginTop: 15,
-    alignItems: 'center'
-  },
-  helpLink: {
-    paddingVertical: 15
-  },
-  helpLinkText: {
-    fontSize: 14,
-    color: '#2e78b7'
-  }
-})
 
 export default withApollo(HomeScreen)
