+++ client/styles/MapScreenStyles.js
@@ -1,0 +1,93 @@
+// import {StyleSheet} from 'react-native'
+
+// const styles = {
+//   container: {
+//     width: '100%',
+//     height: '100%'
+//   },
+//   overlay: {
+//     position: 'absolute',
+//     top: 0
+//   },
+//   // mapContainer: {
+//   //   flex: 1,
+//   //   backgroundColor: '#fff',
+//   //   alignItems: 'center',
+//   //   justifyContent: 'center'
+//   // },
+//   topNav: {
+//     position: 'absolute',
+//     top: 0,
+//     zIndex: 10000000
+//   },
+//   mapStyle: {
+//     width: Dimensions.get('window').width,
+//     height: Dimensions.get('window').height / 2,
+//     marginTop: 0
+//   }
+// }
+
+// const styles2 = StyleSheet.create({
+//   container: {
+//     flex: 1,
+//     backgroundColor: 'transparent',
+//     marginTop: 400,
+//     marginLeft: 190,
+//     position: 'absolute'
+//   },
+//   buttonContainer: {
+//     padding: 5
+//   },
+//   buttonInner: {
+//     marginBottom: 10
+//   },
+//   labelText: {
+//     color: '#333',
+//     fontSize: 12,
+//     fontWeight: '500',
+//     marginBottom: 15
+//   },
+//   item: {
+//     borderWidth: 1,
+//     borderColor: '#333',
+//     backgroundColor: '#FFF'
+//   },
+//   label: {
+//     color: '#333'
+//   },
+//   itemSelected: {
+//     backgroundColor: '#333'
+//   },
+//   labelSelected: {
+//     color: '#FFF'
+//   }
+// })
+
+// const styles99 = StyleSheet.create({
+//   activityIndicatorContainer: {
+//     backgroundColor: '#fff',
+//     alignItems: 'center',
+//     justifyContent: 'center',
+//     flex: 1
+//   },
+
+//   row: {
+//     borderBottomWidth: 1,
+//     borderColor: '#ccc',
+//     padding: 10
+//   },
+
+//   title: {
+//     fontSize: 15,
+//     fontWeight: '600'
+//   },
+
+//   description: {
+//     marginTop: 5,
+//     fontSize: 14
+//   }
+// })
+
+// export default styles  ;
+// export default styles2;
+// export default styles99;
